requirements.txt
pytorch
torchvision
pillow
click
